重构 MockDB 逻辑，简化数据库模拟并添加配置结构；更新性能测试和 HTTP 基准测试以使用新 MockDB
