+++ go/cmd/http_benchmark/main.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"geecache"
 	"geecache/geecachehttp"
+	"geecache/pkg/mockdb"
 	"log"
-	"math/rand"
 	"net/http"
 	"time"
 )
@@ -19,34 +19,13 @@ const (
 	DBDelay  = 10 * time.Millisecond // 模拟 10ms 的 DB 延迟
 )
 
-// MockDB 逻辑
-type MockDB struct {
-	data map[string][]byte
-}
-
-func NewMockDB() *MockDB {
-	db := &MockDB{data: make(map[string][]byte, DBSize)}
-	val := make([]byte, ItemSize)
-	rand.Read(val)
-	for i := 0; i < DBSize; i++ {
-		key := fmt.Sprintf("key_%d", i)
-		db.data[key] = val
-	}
-	return db
-}
-
-func (db *MockDB) Get(key string) ([]byte, error) {
-	// 模拟 DB 延迟
-	time.Sleep(DBDelay)
-	if v, ok := db.data[key]; ok {
-		return v, nil
-	}
-	return nil, fmt.Errorf("not found")
-}
-
 func main() {
 	// 1. 初始化 DB
-	db := NewMockDB()
+	db := mockdb.New(mockdb.Config{
+		Size:      DBSize,
+		ValueSize: ItemSize,
+		Latency:   DBDelay,
+	})
 
 	// 2. 初始化 GeeCache Group
 	// 缓存大小设置大一点，尽量让测试期间能命中
